be sure to not kill the whole session just because somebody breaks KProcess.
svn path=/trunk/kdebase/kcontrol/; revision=206964

+++ kcontrol/ebrowsing/plugins/localdomain/localdomainurifilter.cpp
@@ -112,7 +112,7 @@ bool LocalDomainURIFilter::isLocalDomainHost( const QString& cmd ) const
 	//tm.tv_nsec = 20 * 1000 * 1000; // 20ms
 	usleep( 20000 );
 	}
-    kill( pid, SIGTERM );
+    if ( pid > 0 ) kill( pid, SIGTERM );
     last_result = false;
     return false;
     }
